rename postgres::proto to postgres::handlers
